pyproject.toml
[tool.poetry]
name = "django-journal-dobrynya"
version = "0.1.0"
description = ""
authors = ["vellimir <kkazancov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
django = "^5.1.2"
python-dotenv = "^1.0.1"
django-colorfield = "^0.11.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
